Cleaned up and tested manually. Next step is to add proper testing.
