+++ configurator/src/lib.rs
@@ -1,6 +1,8 @@
-#[cfg(test)]
-use mockall::{automock, mock, predicate::*};
-use mockall_double::double;
+use profile::profile::layout::action::ActionType::{Analog, Digital};
+use profile::profile::layout::Action;
+use profile::profile::Layout;
+use profile::Profile;
+use std::fmt;
 
 mod encoder;
 pub mod profile {
@@ -8,18 +10,69 @@ pub mod profile {
 }
 pub mod profiles;
 
-#[cfg(test)]
-mod tests {
-    use super::*;
+impl fmt::Display for Action {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let action_type = match &self.action_type {
+            Some(x) => x,
+            None => return Err(fmt::Error),
+        };
+        match action_type {
+            Digital(x) => write!(f, "{}", x)?,
+            Analog(x) => write!(f, "{}, {}", x.id, x.value)?,
+        }
+        Ok(())
+    }
+}
 
-    #[test]
-    #[should_panic]
-    fn test_upload_too_many_bytes() {
-        let config = Configurator::default();
-        let encode_ctx = encoder::encode_context();
-        encode_ctx
-            .expect()
-            .return_const(vec![0u8, MAX_EEPROM_BYTES + 1]);
-        config.upload();
+impl fmt::Display for Layout {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let actions = [
+            ("thumb top", self.thumb_top.as_ref()),
+            ("thumb middle", self.thumb_middle.as_ref()),
+            ("thumb bottom", self.thumb_bottom.as_ref()),
+            ("index top", self.index_top.as_ref()),
+            ("index middle", self.index_middle.as_ref()),
+            ("middle top", self.middle_top.as_ref()),
+            ("middle middle", self.middle_middle.as_ref()),
+            ("middle bottom", self.middle_bottom.as_ref()),
+            ("ring top", self.ring_top.as_ref()),
+            ("ring middle", self.ring_middle.as_ref()),
+            ("ring bottom", self.ring_bottom.as_ref()),
+            ("pinky top", self.pinky_top.as_ref()),
+            ("pinky middle", self.pinky_middle.as_ref()),
+            ("pinky bottom", self.pinky_bottom.as_ref()),
+            ("left index extra", self.left_index_extra.as_ref()),
+            ("left middle extra", self.left_middle_extra.as_ref()),
+            ("left ring extra", self.left_ring_extra.as_ref()),
+            ("right index extra", self.right_index_extra.as_ref()),
+            ("right middle extra", self.right_middle_extra.as_ref()),
+            ("right ring extra", self.right_ring_extra.as_ref()),
+        ];
+        for action in actions.iter() {
+            let unwrapped = match action.1 {
+                Some(x) => x,
+                None => return Err(fmt::Error),
+            };
+            writeln!(f, "\t{}: {}", action.0, unwrapped)?;
+        }
+        Ok(())
+    }
+}
+
+impl fmt::Display for Profile {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        writeln!(f, "{{")?;
+        writeln!(f, "\tname: {}", self.name)?;
+        for platform_config in &self.platform_config {
+            writeln!(f, "\tplatform: {}", platform_config.platform)?;
+            writeln!(f, "\tposition: {}", platform_config.position)?;
+        }
+        let layout = match &self.layout {
+            Some(x) => x,
+            None => return Err(fmt::Error),
+        };
+        write!(f, "{}", layout)?;
+        writeln!(f, "}}")?;
+        Ok(())
     }
 }
